Move helper functions to ConvTranspose
